+++ examples/hardware_id_no_lib.rs
@@ -46,10 +46,15 @@ fn main() -> ! {
 
 #[panic_handler]
 fn handle_panic(info: &core::panic::PanicInfo) -> ! {
-    rprintln!("PANIC! {}", info);
-    rprintln!("Location {:?}", info.location());
-    if let Some(pl) = info.payload().downcast_ref::<&str>() {
-        rprintln!("Payload {:?}", pl);
+    rprintln!("PANIC! {}", info.message());
+    if let Some(location) = info.location() {
+        rprintln!(
+            "Panic occurred in file '{}' at line {}",
+            location.file(),
+            location.line(),
+        );
+    } else {
+        rprintln!("Panic occurred but can't get location information...");
     }
     loop {}
 }