+++ src/modules/mod.rs
@@ -1,7 +1,9 @@
 pub mod adc;
 pub mod encoder;
 pub mod gpio;
+pub mod keypad;
 pub mod neopixel;
+pub mod quad_encoder;
 pub mod status;
 pub mod timer;