+++ CHANGELOG.md
@@ -10,15 +10,29 @@ and this project (hopefully) adheres to [Semantic Versioning](https://semver.org
 ### To Be Added
 
 - Digital write functions within the Seesaw GPIO module [PR here](https://github.com/alexeden/adafruit-seesaw/pull/8)
-- Implementation of the Seesaw Keypad module [PR here](https://github.com/alexeden/adafruit-seesaw/pull/6)
 - [Add definition for the NeoDriver meopixel driver board](https://www.adafruit.com/product/5766)
+
+## [0.8.0] - 2025-01-31
+
+### Added
+
+- Add support for the `keypad` Seesaw module
+- Add `NeoTrellis` device definition for the [NeoTrellis 4x4 Keypad](https://www.adafruit.com/product/3954)
+- Add NeoTrellis example similar to the [`ripples.ino` example](https://github.com/adafruit/Adafruit_Seesaw/blob/master/examples/NeoTrellis/ripples/ripples.ino)
+  - [Demo video here](https://storage.googleapis.com/apemedia/neotrellis576.mp4)
+
+## [0.7.0] - 2025-01-15
+
+### Added
+
+- Add `QuadEncoder` module
+- Add device definition for the [QuadRotaryEncoder](https://www.adafruit.com/product/5297) ([@HyperSuperMetaCtrl](https://github.com/HyperSuperMetaCtrl))
 
 ## [0.6.1] - 2024-01-29
 
 ### Changed
 
 - Fix issues caused by extraneous byte written when setting neopixel color [PR here](https://github.com/alexeden/adafruit-seesaw/pull/5)
-
 
 ## [0.6.0] - 2024-01-29